+++ buildkite/annotations_test.go
@@ -1,6 +1,7 @@
 package buildkite
 
 import (
+	"encoding/json"
 	"fmt"
 	"net/http"
 	"reflect"
@@ -57,5 +58,61 @@ func TestAnnotationsService_ListByBuild(t *testing.T) {
 	}
 	if !reflect.DeepEqual(annotations, want) {
 		t.Errorf("ListByBuild returned %+v, want %+v", annotations, want)
+	}
+}
+
+func TestAnnotationsService_Create(t *testing.T) {
+	setup()
+	defer teardown()
+
+	input := &AnnotationCreate{
+		Style:   String("info"),
+		Context: String("default"),
+		Body:    String("<h1>My Markdown Heading</h1>\n<p>An example annotation!</p>"),
+		Append:  Bool(false),
+	}
+
+	mux.HandleFunc("/v2/organizations/my-great-org/pipelines/my-great-pipeline/builds/10/annotations", func(w http.ResponseWriter, r *http.Request) {
+		v := new(AnnotationCreate)
+		json.NewDecoder(r.Body).Decode(&v)
+
+		testMethod(t, r, "POST")
+
+		if !reflect.DeepEqual(v, input) {
+			t.Errorf("Request body = %+v, want %+v", v, input)
+		}
+
+		fmt.Fprint(w,
+			`
+			{
+				"id": "68aef727-f754-48e1-aad8-5f5da8a9960c",
+				"context": "default",
+				"style": "info",
+				"body_html": "<h1>My Markdown Heading</h1>\n<p>An example annotation!</p>",
+				"created_at": "2023-08-21T08:50:05.824Z",
+				"updated_at": "2023-08-21T08:50:05.824Z"
+			}`)
+	})
+
+	annotation, _, err := client.Annotations.Create("my-great-org", "my-great-pipeline", "10", input)
+
+	if err != nil {
+		t.Errorf("TestAnnotations.Create returned error: %v", err)
+	}
+
+	annotationCreatedAt, err := time.Parse(BuildKiteDateFormat, "2023-08-21T08:50:05.824Z")
+	annotationUpatedAt, err := time.Parse(BuildKiteDateFormat, "2023-08-21T08:50:05.824Z")
+
+	want := &Annotation{
+		ID:        String("68aef727-f754-48e1-aad8-5f5da8a9960c"),
+		Context:   String("default"),
+		Style:     String("info"),
+		BodyHTML:  String("<h1>My Markdown Heading</h1>\n<p>An example annotation!</p>"),
+		CreatedAt: NewTimestamp(annotationCreatedAt),
+		UpdatedAt: NewTimestamp(annotationUpatedAt),
+	}
+
+	if !reflect.DeepEqual(annotation, want) {
+		t.Errorf("TestAnnotations.Create returned %+v, want %+v", annotation, want)
 	}
 }